+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 ## Changelog
 
-### Version 1.8.7
+### Version 1.8.11
 
+- Project Analysis feature added showing Javascript and Asset sizes
+- Capacitor 3 to 4 migration feature added
+- Recommend install of @awesome-cordova-plugins/core if using a plugin
+- Recommend @capacitor/google-maps on Capacitor migration where cordova-plugin-googlemaps is present
+- Add generated splash and icon resources to .gitignore when generating for Capacitor
 - Fix for retrying builds when a bug is fixed in code and file saved
+- Fix incorrect migration of cordova-plugin-push
 
 ### Version 1.8.6